refactoring * edit task manager to integrate same task name's data
